Qual: Uniformize SQL code to be more compatible

+++ htdocs/install/mysql/tables/llx_commandedet.sql
@@ -41,7 +41,7 @@ create table llx_commandedet
   info_bits          integer      DEFAULT 0,          -- TVA NPR ou non
   marge_tx           double(6,3)  DEFAULT 0,          -- taux de marge (marge sur prix d'achat)
   marque_tx          double(6,3)  DEFAULT 0,          -- taux de marque (marge sur prix de vente)
-  special_code       tinyint(4)   UNSIGNED DEFAULT 0, -- code pour les lignes speciales
+  special_code       integer UNSIGNED DEFAULT 0,      -- code pour les lignes speciales
   rang               integer      DEFAULT 0,
   import_key         varchar(14)
 )type=innodb;
