+++ htdocs/install/mysql/tables/llx_menu.sql
@@ -35,7 +35,7 @@ CREATE TABLE llx_menu
 	target        varchar(100) NULL,                -- Target of Url link
 	titre         varchar(255) NOT NULL,            -- Key for menu translation 
 	langs         varchar(100),                     -- Lang file to load for translation
-	level         tinyint(1),                       -- Used by auguria menu only. Do not use.
+	level         smallint,                         -- Used by auguria menu only. Do not use.
 	leftmenu      varchar(1) NULL default '1',      -- Say if left menu defined in pre.inc.php and used by top menu must be overwritten by dynamic databse menu (1=yes by default)
 	perms         varchar(255),                     -- Condition to show enabled or disabled
 	enabled       varchar(255) NULL default '1',    -- Condition to show or hide
